Move logic from Vm ot service class to reuse it in API
